[test] apply the new test style to common_functions
