+++ common/functions/src/aggregates/mod.rs
@@ -12,11 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod aggregate_combinator_test;
-#[cfg(test)]
-mod aggregate_function_test;
-
 mod aggregate_arg_min_max;
 mod aggregate_avg;
 mod aggregate_combinator_distinct;
