+++ common/functions/src/scalars/hashes/mod.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod siphash_test;
-
 mod hash;
 mod siphash;
