+++ common/functions/src/scalars/logics/mod.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod logic_test;
-
 mod logic;
 mod logic_and;
 mod logic_not;
