+++ common/functions/src/scalars/others/mod.rs
@@ -11,10 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#[cfg(test)]
-mod running_difference_function_test;
 
 mod other;
 mod running_difference_function;
+
 pub use other::OtherFunction;
 pub use running_difference_function::RunningDifferenceFunction;
