+++ common/functions/src/scalars/strings/mod.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod substring_test;
-
 mod string;
 mod substring;
