+++ common/functions/tests/it/aggregates/mod.rs
@@ -12,8 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod conditional;
-mod r#if;
-
-pub use conditional::ConditionalFunction;
-pub use r#if::IfFunction;
+mod aggregate_combinator;
+mod aggregate_function;
