+++ common/functions/tests/it/scalars/arithmetic.rs
@@ -15,9 +15,8 @@
 use common_datavalues::chrono;
 use common_datavalues::prelude::*;
 use common_exception::Result;
+use common_functions::scalars::*;
 use pretty_assertions::assert_eq;
-
-use crate::scalars::*;
 
 #[test]
 fn test_arithmetic_function() -> Result<()> {
