+++ common/functions/tests/it/scalars/dates/date.rs
@@ -11,12 +11,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 use common_datavalues::prelude::*;
 use common_exception::Result;
+use common_functions::scalars::*;
 use pretty_assertions::assert_eq;
-
-use super::RoundFunction;
-use crate::scalars::*;
 
 #[allow(dead_code)]
 struct Test {
