+++ common/functions/tests/it/scalars/expression.rs
@@ -14,9 +14,8 @@
 
 use common_datavalues::prelude::*;
 use common_exception::Result;
+use common_functions::scalars::*;
 use pretty_assertions::assert_eq;
-
-use crate::scalars::*;
 
 #[test]
 fn test_cast_function() -> Result<()> {
