+++ common/functions/tests/it/scalars/function_column.rs
@@ -15,8 +15,7 @@
 use common_datablocks::*;
 use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::scalars::*;
+use common_functions::scalars::*;
 
 #[test]
 fn test_column_function() -> Result<()> {
