+++ common/functions/tests/it/scalars/maths/mod.rs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod conditional;
-mod r#if;
-
-pub use conditional::ConditionalFunction;
-pub use r#if::IfFunction;
+mod abs;
+mod crc32;
+mod pi;
+mod trigonometric;
