+++ common/functions/tests/it/scalars/maths/trigonometric.rs
@@ -16,8 +16,7 @@ use std::f64::consts::PI;
 
 use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::scalars::*;
+use common_functions::scalars::*;
 
 #[test]
 fn test_trigonometic_function() -> Result<()> {
