+++ common/functions/tests/it/scalars/mod.rs
@@ -12,8 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod arithmetic;
+mod comparison;
 mod conditional;
-mod r#if;
-
-pub use conditional::ConditionalFunction;
-pub use r#if::IfFunction;
+mod dates;
+mod expression;
+mod function_column;
+mod hash;
+mod logic;
+mod maths;
+mod nullable;
+mod other;
+mod udfs;
