+++ common/functions/tests/it/scalars/udfs/mod.rs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod conditional;
-mod r#if;
-
-pub use conditional::ConditionalFunction;
-pub use r#if::IfFunction;
+mod database;
+mod to_type_name;
+mod udf_example;
+mod version;
