+++ common/functions/tests/it/scalars/udfs/udf_example.rs
@@ -12,59 +12,51 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datavalues::columns::DataColumn;
 use common_datavalues::prelude::*;
+use common_datavalues::DataField;
+use common_datavalues::DataSchemaRefExt;
 use common_datavalues::DataType;
 use common_exception::Result;
-use pretty_assertions::assert_eq;
-
-use crate::scalars::*;
+use common_functions::scalars::*;
 
 #[test]
-fn test_if_function() -> Result<()> {
+fn test_udf_example_function() -> Result<()> {
     #[allow(dead_code)]
     struct Test {
         name: &'static str,
         display: &'static str,
         nullable: bool,
-        args: Vec<DataType>,
         columns: Vec<DataColumn>,
-        expect: Series,
+        expect: DataColumn,
         error: &'static str,
         func: Box<dyn Function>,
     }
 
     let schema = DataSchemaRefExt::create(vec![
-        DataField::new("a", DataType::Int32, false),
-        DataField::new("b", DataType::Int64, false),
+        DataField::new("a", DataType::Boolean, false),
+        DataField::new("b", DataType::Boolean, false),
     ]);
 
     let tests = vec![Test {
-        name: "if-passed",
-        display: "IF",
+        name: "udf-example-passed",
+        display: "example()",
         nullable: false,
-        func: IfFunction::try_create_func("")?,
-        args: vec![
-            DataType::Boolean,
-            schema.field_with_name("a")?.data_type().clone(),
-            DataType::Float64,
-        ],
+        func: UdfExampleFunction::try_create("example")?,
         columns: vec![
-            Series::new(vec![true, false, false, true]).into(),
-            Series::new(vec![1i32, 2, 3, 4]).into(),
-            DataColumn::Constant(DataValue::Float64(Some(2.5)), 4),
+            Series::new(vec![true, true, true, false]).into(),
+            Series::new(vec![true, false, true, true]).into(),
         ],
-        expect: Series::new(vec![1f64, 2.5, 2.5, 4f64]),
+        expect: Series::new(vec![true, true, true, true]).into(),
         error: "",
     }];
 
     for t in tests {
         let func = t.func;
 
-        let columns: Vec<DataColumnWithField> = t
-            .columns
-            .iter()
-            .map(|c| DataColumnWithField::new(c.clone(), DataField::new("a", c.data_type(), false)))
-            .collect();
+        if let Err(e) = func.eval(&[], t.columns[0].len()) {
+            assert_eq!(t.error, e.to_string());
+        }
 
         // Display check.
         let expect_display = t.display.to_string();
@@ -76,15 +68,12 @@ fn test_if_function() -> Result<()> {
         let actual_null = func.nullable(&schema)?;
         assert_eq!(expect_null, actual_null);
 
+        let columns = vec![];
         let v = &(func.eval(&columns, t.columns[0].len())?);
-        // Type check.
-        let expect_type = func.return_type(&t.args)?;
+        let expect_type = func.return_type(&[])?;
         let actual_type = v.data_type();
         assert_eq!(expect_type, actual_type);
-
-        let cmp = v.to_array()?.eq(&t.expect)?;
-        assert!(cmp.all_true());
+        assert_eq!(v, &t.expect);
     }
-
     Ok(())
 }
